backend/requirements.txt
opencv-python==4.13.0.92
mediapipe==0.10.33
librosa==0.11.0
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.26
openai-whisper==20250625
